serverless/requirements.txt
boto3==1.9.216
ask-sdk-core==1.11.0
ask-sdk
darksky_weather==1.5.0
geopy==1.20.0
